bbx-sdk/requirements.txt
# BBX Python SDK Requirements

httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0
